requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
pytesseract>=0.3.8
matplotlib>=3.5.0
tqdm>=4.60.0
Pillow>=8.0.0
pandas>=1.3.0
torch>=1.7.0
torchvision>=0.8.0
pyyaml>=5.4.0 